Compile fixes

+++ src/doom/r_draw.cpp
@@ -745,9 +745,9 @@ void R_InitTranslationTables()
         if (i >= 0x70 && i <= 0x7f)
         {
             // map green ramp to gray, brown, red
-            translationtables[i]       = 0x60 + (i & 0xf);
-            translationtables[i + 256] = 0x40 + (i & 0xf);
-            translationtables[i + 512] = 0x20 + (i & 0xf);
+            translationtables[i]       = static_cast<uint8_t>(0x60 + (i & 0xf));
+            translationtables[i + 256] = static_cast<uint8_t>(0x40 + (i & 0xf));
+            translationtables[i + 512] = static_cast<uint8_t>(0x20 + (i & 0xf));
         }
         else
         {
