+++ src/strife/r_draw.cpp
@@ -467,13 +467,13 @@ void R_InitTranslationTables ()
     {
         if(i >= 0x80 && i <= 0x8f)
         {
-            translationtables [i      ] = (i & 0x0f) + 64;
-            translationtables [i+  256] = (i & 0x0f) - 80;
-            translationtables [i+2*256] = (i & 0x0f) + 16;
-            translationtables [i+3*256] = (i & 0x0f) + 48;
-            translationtables [i+4*256] = (i & 0x0f) + 80;
-            translationtables [i+5*256] = (i & 0x0f) + 96;
-            translationtables [i+6*256] = (i & 0x0f) - 112;
+            translationtables [i      ] = static_cast<uint8_t>((i & 0x0f) + 64);
+            translationtables [i+  256] = static_cast<uint8_t>((i & 0x0f) - 80);
+            translationtables [i+2*256] = static_cast<uint8_t>((i & 0x0f) + 16);
+            translationtables [i+3*256] = static_cast<uint8_t>((i & 0x0f) + 48);
+            translationtables [i+4*256] = static_cast<uint8_t>((i & 0x0f) + 80);
+            translationtables [i+5*256] = static_cast<uint8_t>((i & 0x0f) + 96);
+            translationtables [i+6*256] = static_cast<uint8_t>((i & 0x0f) - 112);
         }
         else if(i >= 0x50 && i<= 0x5f)
         {
@@ -481,9 +481,9 @@ void R_InitTranslationTables ()
             translationtables [i+  256] = static_cast<uint8_t>(i);
             translationtables [i+2*256] = static_cast<uint8_t>(i);
             translationtables [i+3*256] = static_cast<uint8_t>(i);
-            translationtables [i+4*256] = (i & 0x0f) + 0x80;
-            translationtables [i+5*256] = (i & 0x0f) + 16;
-            translationtables [i+6*256] = (i & 0x0f) + 64;
+            translationtables [i+4*256] = static_cast<uint8_t>((i & 0x0f) + 0x80);
+            translationtables [i+5*256] = static_cast<uint8_t>((i & 0x0f) + 16);
+            translationtables [i+6*256] = static_cast<uint8_t>((i & 0x0f) + 64);
         }
         else if(i >= 0xd0 && i<= 0xdf)
         {
@@ -491,9 +491,9 @@ void R_InitTranslationTables ()
             translationtables [i+  256] = static_cast<uint8_t>(i);
             translationtables [i+2*256] = static_cast<uint8_t>(i);
             translationtables [i+3*256] = static_cast<uint8_t>(i);
-            translationtables [i+4*256] = (i & 0x0f) - 80;
-            translationtables [i+5*256] = (i & 0x0f) + 48;
-            translationtables [i+6*256] = (i & 0x0f) + 16;
+            translationtables [i+4*256] = static_cast<uint8_t>((i & 0x0f) - 80);
+            translationtables [i+5*256] = static_cast<uint8_t>((i & 0x0f) + 48);
+            translationtables [i+6*256] = static_cast<uint8_t>((i & 0x0f) + 16);
         }
         else if(i >= 0xc0 && i<= 0xcf)
         {
@@ -501,8 +501,8 @@ void R_InitTranslationTables ()
             translationtables [i+  256] = static_cast<uint8_t>(i);
             translationtables [i+2*256] = static_cast<uint8_t>(i);
             translationtables [i+3*256] = static_cast<uint8_t>(i);
-            translationtables [i+4*256] = (i & 0x0f) - 96;
-            translationtables [i+5*256] = (i & 0x0f) + 32;
+            translationtables [i+4*256] = static_cast<uint8_t>((i & 0x0f) - 96);
+            translationtables [i+5*256] = static_cast<uint8_t>((i & 0x0f) + 32);
             translationtables [i+6*256] = (i & 0x0f);
             // haleyjd 20110213: missing code:
             if(!(i & 0x0f))
@@ -520,7 +520,7 @@ void R_InitTranslationTables ()
         }
         else if(i >= 0xf1 && i<= 0xf6)
         {
-            translationtables [i      ] = (i & 0x0f) - 33;
+            translationtables [i      ] = static_cast<uint8_t>((i & 0x0f) - 33);
             translationtables [i+  256] = static_cast<uint8_t>(i);
             translationtables [i+2*256] = static_cast<uint8_t>(i);
             translationtables [i+3*256] = static_cast<uint8_t>(i);
@@ -532,8 +532,8 @@ void R_InitTranslationTables ()
         {
             translationtables [i      ] = col2;
             translationtables [i+  256] = col2;
-            translationtables [i+2*256] = (i & 0x0f) - 48;
-            translationtables [i+3*256] = (i & 0x0f) - 48;
+            translationtables [i+2*256] = static_cast<uint8_t>((i & 0x0f) - 48);
+            translationtables [i+3*256] = static_cast<uint8_t>((i & 0x0f) - 48);
             translationtables [i+4*256] = col2;
             translationtables [i+5*256] = col2;
             translationtables [i+6*256] = col2;
